Correct nginx example
The "proxy_protocol" in the listen directive refers to incoming connections,
that is, if you want to put nginx behind another proxy.

This results in errors like: "broken header: "<binary data>" while reading PROXY protocol".

Only "proxy_protocol on" is needed to make it generate those headers and send them
to stalwart.